Backport from trunk rev7811: Bugfix(#328): synchronisation error when using Active directory User
git-svn-id: file:///home/code/svn/ovd/branches/3.0@7817 2b2718be-6ee5-4ec5-ad1a-3fa14f9c7ca3

+++ OvdServer/ovd/Role/ApplicationServer/Session.py
@@ -4,7 +4,7 @@
 # http://www.ulteo.com
 # Author Laurent CLOUET <laurent@ulteo.com> 2010
 # Author Julien LANGLOIS <julien@ulteo.com> 2009-2010
-# Author David LECHEVALIER <david@ulteo.com> 2010
+# Author David LECHEVALIER <david@ulteo.com> 2010, 2011
 #
 # This program is free software; you can redistribute it and/or 
 # modify it under the terms of the GNU General Public License
@@ -261,6 +261,7 @@ class Session:
 			shutil.copyfile(path, dst)
 			
 			try:
-				os.remove(path)
+				if self.domain.manage_user():
+					os.remove(path)
 			except:
 				pass
